# contracts/http/platform-api.openapi.yaml
openapi: 3.1.0
info:
  title: Backend Architecture Learning Platform API
  version: 0.3.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
security: []
...
paths:
...
  /api/v1/courses:
    get:
      tags: [Courses]
      summary: List knowledge-map courses
      operationId: listCourses
      security:
        - sessionCookie: []
        - {}
      responses:
        "200":
          description: Knowledge-map courses in display order with optional user progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/courses/{slug}:
    get:
      tags: [Courses]
      summary: Get one course and record an authenticated user's view
      operationId: getCourse
      security:
        - sessionCookie: []
        - {}
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9-]{0,127}$"
      responses:
        "200":
          description: Course metadata, theory content, lab availability, and optional progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetailResponse"
        "404":
          description: Course does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/labs:
    post:
      tags: [Labs]
      summary: Create a lab session
      operationId: createLab
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabCreateRequest"
      responses:
        "202":
          description: The lab session and its idempotent create operation were accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabCreateResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/CSRFInvalid"
        "404":
          description: The requested course does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The user already has an active lab or the operation conflicts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Lab capacity or trusted scenario resources are unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/labs/{labId}:
    get:
      tags: [Labs]
      summary: Get an owned lab snapshot
      operationId: getLab
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/LabID"
      responses:
        "200":
          description: Current persisted lab, topology, and latest operation state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabSnapshotResponse"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "404":
          description: The lab does not exist or belongs to another user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Labs]
      summary: Terminate and destroy an owned lab
      operationId: terminateLab
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/LabID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOperationRequest"
      responses:
        "202":
          description: The idempotent destroy operation was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabActionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/CSRFInvalid"
        "404":
          description: The lab does not exist or belongs to another user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The lab is busy or cannot be terminated from its current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/labs/{labId}/reset:
    post:
      tags: [Labs]
      summary: Reset an owned lab to its initial scenario state
      operationId: resetLab
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/LabID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabOperationRequest"
      responses:
        "202":
          description: The idempotent reset operation was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabActionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/CSRFInvalid"
        "404":
          description: The lab does not exist or belongs to another user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The lab is busy or not in a resettable running state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/labs/{labId}/traffic-batches:
    post:
      tags: [Labs]
      summary: Submit one frontend-generated traffic batch
      operationId: submitTrafficBatch
      security:
        - sessionCookie: []
      parameters:
        - name: labId
          in: path
          required: true
          schema:
            type: string
            pattern: "^[a-z0-9][a-z0-9-]{3,63}$"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrafficBatchRequest"
      responses:
        "200":
          description: Batch was processed, partially processed, or dropped by capacity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrafficBatchResponse"
        "400":
          description: Batch fields, product, or request units are invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/CSRFInvalid"
        "404":
          description: The lab does not exist or is not owned by the user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The lab is not running.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: The platform HTTP rate limit was exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: The lab gateway, instance, or internal response is unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/labs/{labId}/actions:
    post:
      tags: [Labs]
      summary: Submit one guided cluster topology action
      operationId: submitLabAction
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/LabID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTopologyActionRequest"
      responses:
        "202":
          description: The idempotent topology action was accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabActionResponse"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/AuthRequired"
        "403":
          $ref: "#/components/responses/CSRFInvalid"
        "404":
          description: The lab or target instance does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: The lab is busy, not running, or at an instance boundary.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: The action is not available for the current scenario or mode.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Trusted capacity, Docker, or Nginx resources are unavailable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: HttpOnly Session Cookie set by the login endpoint.
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: CSRF token returned by login or the current-user endpoint.
      schema:
        type: string
        minLength: 1
    LabID:
      name: labId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9][a-z0-9-]{3,63}$"
  responses:
    ValidationFailed:
      description: Request JSON or fields are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    AuthRequired:
      description: A valid Session Cookie is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    CSRFInvalid:
      description: The CSRF token is missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: The request could not be completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CourseProgress:
      type: object
      required: [viewed, lastViewedAt, lastLabId]
      properties:
        viewed:
          type: boolean
        lastViewedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        lastLabId:
          oneOf:
            - type: string
            - type: "null"
    Course:
      type: object
      required:
        - id
        - slug
        - title
        - category
        - status
        - sortOrder
        - summary
        - labAvailable
        - progress
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
        title:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [theory, active, coming_soon]
        sortOrder:
          type: integer
        summary:
          type: string
        labAvailable:
          type: boolean
        progress:
          oneOf:
            - $ref: "#/components/schemas/CourseProgress"
            - type: "null"
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          required:
            - content
            - contentFormat
            - implementation
            - lab
          properties:
            content:
              type: string
            contentFormat:
              type: string
              const: markdown
            implementation:
              type: object
              required: [requestPath, keyConcepts]
              properties:
                requestPath:
                  type: array
                  items:
                    type: string
                keyConcepts:
                  type: array
                  items:
                    type: string
            lab:
              type: object
              required: [available]
              properties:
                available:
                  type: boolean
                scenarioType:
                  type: string
                allowedInstanceRange:
                  type: object
                  required: [min, max]
                  properties:
                    min:
                      type: integer
                    max:
                      type: integer
    CourseListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [courses]
          properties:
            courses:
              type: array
              items:
                $ref: "#/components/schemas/Course"
        meta:
          type: object
          required: [authenticated, total]
          properties:
            authenticated:
              type: boolean
            total:
              type: integer
    CourseDetailResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [course]
          properties:
            course:
              $ref: "#/components/schemas/CourseDetail"
    LabCreateRequest:
      type: object
      additionalProperties: false
      required: [operationId, courseId]
      properties:
        operationId:
          type: string
          minLength: 1
          maxLength: 64
        courseId:
          type: integer
          format: int64
          minimum: 1
    LabOperationRequest:
      type: object
      additionalProperties: false
      required: [operationId]
      properties:
        operationId:
          type: string
          minLength: 1
          maxLength: 64
    LabTopologyActionRequest:
      type: object
      additionalProperties: false
      required: [operationId, actionType, targetInstanceId, parameters]
      properties:
        operationId:
          type: string
          minLength: 1
          maxLength: 64
        actionType:
          type: string
          enum: [ADD_INSTANCE, REMOVE_INSTANCE, SET_INSTANCE_PERFORMANCE, SET_INSTANCE_WEIGHTS]
        targetInstanceId:
          oneOf:
            - type: string
            - type: "null"
        parameters:
          type: object
          additionalProperties: false
          properties:
            performancePercent:
              type: integer
              minimum: 20
              maximum: 100
            weights:
              type: array
              items:
                type: object
                additionalProperties: false
                required: [instanceId, weight]
                properties:
                  instanceId:
                    type: string
                  weight:
                    type: integer
                    minimum: 1
                    maximum: 100
    LabSession:
      type: object
      required:
        - id
        - courseId
        - scenarioType
        - scenarioTemplateId
        - status
        - balancingMode
        - redisEnabled
        - startedAt
        - lastEffectiveActionAt
        - idleExpiresAt
        - maximumExpiresAt
        - terminatedAt
        - terminationReason
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        courseId:
          type: integer
          format: int64
        scenarioType:
          type: string
        scenarioTemplateId:
          type: string
        status:
          type: string
          enum: [Preparing, Running, Expiring, Failed, Terminating, Terminated]
        balancingMode:
          type: string
          enum: [fixed, adaptive]
        redisEnabled:
          type: boolean
        startedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        lastEffectiveActionAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        idleExpiresAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        maximumExpiresAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        terminatedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        terminationReason:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LabOperation:
      type: object
      required: [operationId, labId, action, status, submittedAt]
      properties:
        operationId:
          type: string
        labId:
          type: string
        action:
          type: string
          enum: [CREATE_LAB, RESET_LAB, DESTROY_LAB]
        status:
          type: string
          enum: [pending, claimed, running, succeeded, failed, compensating]
        submittedAt:
          type: string
          format: date-time
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
    LabInstance:
      type: object
      required:
        - instanceId
        - instanceName
        - status
        - cpuLimitCores
        - memoryLimitMb
        - performancePercent
        - effectiveCapacity
        - currentWeight
        - remainingCapacity
        - loadRatio
        - loadState
      properties:
        instanceId:
          type: string
        instanceName:
          type: string
        status:
          type: string
        cpuLimitCores:
          type: number
        memoryLimitMb:
          type: integer
        performancePercent:
          type: integer
        effectiveCapacity:
          type: integer
        currentWeight:
          type: integer
        remainingCapacity:
          type: integer
          minimum: 0
        loadRatio:
          type: number
          minimum: 0
        loadState:
          type: string
          enum: [idle, normal, high, overloaded, unavailable]
    LabSnapshotResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [lab, topology, latestOperation, trafficPolicy]
          properties:
            lab:
              $ref: "#/components/schemas/LabSession"
            topology:
              type: object
              required: [instances, redis, gateway, nodes, edges]
              properties:
                instances:
                  type: array
                  items:
                    $ref: "#/components/schemas/LabInstance"
                redis:
                  oneOf:
                    - type: object
                      required: [status]
                      properties:
                        status:
                          type: string
                    - type: "null"
                gateway:
                  type: object
                  required: [status]
                  properties:
                    status:
                      type: string
                nodes:
                  type: array
                  items:
                    type: object
                    required: [id, type]
                    properties:
                      id:
                        type: string
                      type:
                        type: string
                        enum: [user_pool, gateway, application]
                edges:
                  type: array
                  items:
                    type: object
                    required: [from, to]
                    properties:
                      from:
                        type: string
                      to:
                        type: string
            latestOperation:
              oneOf:
                - $ref: "#/components/schemas/LabOperation"
                - type: "null"
            trafficPolicy:
              type: object
              required: [requestUnits, generationIntervalMs]
              properties:
                requestUnits:
                  $ref: "#/components/schemas/IntegerPolicy"
                generationIntervalMs:
                  $ref: "#/components/schemas/IntegerPolicy"
    IntegerPolicy:
      type: object
      required: [minimum, maximum, step, default]
      properties:
        minimum:
          type: integer
        maximum:
          type: integer
        step:
          type: integer
        default:
          type: integer
    LabCreateResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [lab, operation]
          properties:
            lab:
              $ref: "#/components/schemas/LabSession"
            operation:
              $ref: "#/components/schemas/LabOperation"
    LabActionResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [operation]
          properties:
            operation:
              $ref: "#/components/schemas/LabOperation"
    TrafficBatchRequest:
      type: object
      additionalProperties: false
      required: [batchId, productId, requestUnits]
      properties:
        batchId:
          type: string
          minLength: 1
          maxLength: 64
        productId:
          type: integer
          format: int64
          minimum: 1
        requestUnits:
          type: integer
          minimum: 1
          maximum: 100
    TrafficBatchResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [result]
          properties:
            result:
              $ref: "#/components/schemas/TrafficBatchResult"
    TrafficBatchResult:
      type: object
      required:
        - batchId
        - labId
        - status
        - occurredAt
        - path
        - targetInstanceId
        - receivedUnits
        - processedUnits
        - droppedUnits
        - instanceState
      properties:
        batchId:
          type: string
        labId:
          type: string
        status:
          type: string
          enum: [processed, partially_processed, dropped]
        occurredAt:
          type: string
          format: date-time
        path:
          type: array
          minItems: 2
          items:
            type: string
        targetInstanceId:
          type: string
        receivedUnits:
          type: integer
          minimum: 0
        processedUnits:
          type: integer
          minimum: 0
        droppedUnits:
          type: integer
          minimum: 0
        instanceState:
          $ref: "#/components/schemas/TrafficInstanceState"
    TrafficInstanceState:
      type: object
      required:
        - effectiveCapacity
        - remainingCapacity
        - loadRatio
        - loadState
        - capacityWindowStartedAt
        - capacityWindowEndsAt
      properties:
        effectiveCapacity:
          type: integer
          minimum: 0
        remainingCapacity:
          type: integer
          minimum: 0
        loadRatio:
          type: number
          minimum: 0
        loadState:
          type: string
          enum: [idle, normal, high, overloaded, unavailable]
        capacityWindowStartedAt:
          type: string
          format: date-time
        capacityWindowEndsAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        requestId:
          type: string